Improve quitting behavior
Introduce “q” as an alias for “quit”, and document them. Also make
Ctrl+D work. Also warn the user if there are legitimate “q” or “quit”
files in the directory; and also rearrange the checks so that, in that
case, guessing these files *correctly* still works (but if you get it
wrong and guess “q” or “quit” when it was something else, then it’s
treated as quitting rather than a wrong guess).