Replay music after wrong guess
I’m hoping that listening to it again, knowing what it is this time,
will help the user to remember it.

I tried extracting the ffplay command into a function, but it didn’t
work out (it no longer seemed to be killed after the user entered their
guess), so let’s just copy+paste it.